Remove fixtures
