--- src/__fixtures__/basic.js
@@ -1,6 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-defineMessages({
-  hello: 'hello',
-  world: 'hello world',
-})
