--- src/__fixtures__/import-as.js
@@ -1,6 +1,0 @@
-import { defineMessages as m } from 'react-intl'
-
-m({
-  hello: 'hello',
-  world: 'hello world',
-})
