--- src/__fixtures__/multi-export.js
@@ -1,11 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-export const extra = defineMessages({
-  hello: 'hello extra',
-  world: 'hello world extra',
-})
-
-export default defineMessages({
-  hello: 'hello',
-  world: 'hello world',
-})
