--- src/__fixtures__/string-literal.js
@@ -1,6 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-defineMessages({
-  'hello': 'hello', // prettier-ignore
-  'world': 'hello world', // prettier-ignore
-})
