--- src/__fixtures__/with-other-func.js
@@ -1,10 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-defineMessages({
-  hello: 'hello',
-  world: `hello world ${1}`,
-})
-
-hello({
-  id: 'hoge',
-})
